Move checker parser to do_if package
